python-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
supabase==2.3.4
requests==2.31.0
python-dotenv==1.0.0

# Optional dependencies for video processing
# Uncomment based on your processing needs:

# For OpenCV video processing
# opencv-python==4.9.0.80

# For FFmpeg integration
# ffmpeg-python==0.2.0

# For machine learning models
# torch==2.1.2
# torchvision==0.16.2
# transformers==4.36.2

# For additional image/video processing
# Pillow==10.1.0
# numpy==1.24.3
# scipy==1.11.4